socketpair: re-implement using Unix domain socket code
Beside reducing code duplication, this change fixes a logical issue
with the previous socketpair implementation, in that a socket pair
was using a single data buffer shared between the two sockets (so
an application could read from a socket the data that had been
written to the same socket), whereas in a correct implementation
each socket in a pair has its own buffered data, and data witten to
one socket can only be read from the other socket.
